requirements.txt
anyio==3.3.4; python_full_version >= '3.6.2'
asgiref==3.4.1; python_version >= '3.6'
certifi==2021.10.8
charset-normalizer==2.0.7; python_version >= '3'
click==8.0.3; python_version >= '3.6'
fastapi-utils==0.2.1; python_version >= '3.6' and python_version < '4'
fastapi==0.70.0
git+https://github.com/udaan-com/udaan-py-common.git@2bbde409c8ea7020624d83a274f877982cbf91f6#egg=udaan-py-common
greenlet==1.1.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32'))))) and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.12.0; python_version >= '3.6'
httptools==0.2.0
idna==3.3; python_version >= '3'
importlib-metadata==4.8.2; python_version < '3.8'
numpy==1.19.5
pandas==1.1.5
pillow==8.4.0
prometheus-client==0.12.0
pydantic==1.8.2
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
pyyaml==6.0
requests==2.26.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
sqlalchemy==1.4.27; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
starlette-exporter==0.11.0
starlette==0.16.0
typing-extensions==4.0.0; python_version < '3.8'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn[standard]==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
wheel==0.36.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
zipp==3.6.0; python_version >= '3.6'
open3d
omegaconf
pytorch-lightning
mlflow
py3rosmsgs
bagpy